Move out read_unsigned_varint
