+++ src/varint.rs
@@ -1,3 +1,5 @@
+use crate::slice_reader::SliceReader;
+
 #[allow(dead_code)]
 #[inline]
 pub fn to_zigzag32(n: i32) -> u32 {
@@ -20,6 +22,25 @@ pub fn to_zigzag64(n: i64) -> u64 {
 #[inline]
 pub fn from_zigzag64(n: u64) -> i64 {
     ((n >> 1) as i64) ^ (-((n & 1) as i64))
+}
+
+pub fn read_unsigned_varint(data: &mut SliceReader) -> Option<u64> {
+    let mut out = 0u64;
+    let mut byte_counter = 0;
+    loop {
+        let Some(byte) = data.read_one() else {
+            return None;
+        };
+        let value = (byte & 0x7f) as u64;
+        out += value << (byte_counter * 7);
+
+        byte_counter += 1;
+        if byte & 0x80 == 0 {
+            break;
+        }
+    }
+
+    Some(out)
 }
 
 #[cfg(test)]
@@ -152,6 +173,53 @@ mod tests {
             let n = (0b1000000000000000000000000000000000000000000000000000000000000000u64 as i64)
                 | (1 << i);
             assert_eq!(from_zigzag64(to_zigzag64(n)), n);
+        }
+    }
+
+    #[test]
+    fn read_unsigned_varint_works() {
+        // tests from https://codeberg.org/ft/ufw/src/tag/v4.1.0/test/t-varint.c#L92-L101
+        {
+            let original = vec![0u8];
+            let mut reader = SliceReader::new(&original);
+            assert_eq!(read_unsigned_varint(&mut reader).unwrap(), 0);
+            assert_eq!(reader.len(), 0);
+
+            let original = vec![0x80, 0x01];
+            let mut reader = SliceReader::new(&original);
+            assert_eq!(read_unsigned_varint(&mut reader).unwrap(), 128);
+            assert_eq!(reader.len(), 0);
+
+            let original = vec![0xd2, 0x09];
+            let mut reader = SliceReader::new(&original);
+            assert_eq!(read_unsigned_varint(&mut reader).unwrap(), 1234);
+            assert_eq!(reader.len(), 0);
+
+            let original = vec![0xff, 0xff, 0xff, 0xff, 0x0f];
+            let mut reader = SliceReader::new(&original);
+            assert_eq!(read_unsigned_varint(&mut reader).unwrap(), u32::MAX as u64);
+            assert_eq!(reader.len(), 0);
+        }
+
+        // examples from https://github.com/multiformats/unsigned-varint
+        {
+            // two different representations of the same value
+            let mut reader = SliceReader::new(&[0x81, 0x00]);
+            assert_eq!(read_unsigned_varint(&mut reader).unwrap(), 1);
+            let mut reader = SliceReader::new(&[0x1]);
+            assert_eq!(read_unsigned_varint(&mut reader).unwrap(), 1);
+
+            for (expected, data) in [
+                (1, vec![0x01]),
+                (127, vec![0x7f]),
+                (128, vec![0x80, 0x01]),
+                (255, vec![0xff, 0x01]),
+                (300, vec![0xac, 0x02]),
+                (16384, vec![0x80, 0x80, 0x01]),
+            ] {
+                let mut reader = SliceReader::new(&data);
+                assert_eq!(read_unsigned_varint(&mut reader).unwrap(), expected);
+            }
         }
     }
 }
